django-template/requirements.txt
# django base
http://www.djangoproject.com/download/1.3-beta-1/tarball/

# database migration
south==0.7.3

# development helpers
django-debug-toolbar==0.8.4
django-sentry==1.6.0
